training_pipeline/requirements.txt
python-dotenv==1.0.0
gcsfs==2023.6.0
fsspec==2023.6.0
google-cloud-storage==2.7.0
numpy==1.23.5
pandas==1.5.3
joblib==1.2.0
category_encoders==2.6.1
scikit-learn==1.2.1
sqlalchemy==1.4.46